Redirect security-credentials requests in the metaproxy (#3)
The AWS Metadata service used to have a behavior that when
/meta-data/iam/security-credentials (without an ending slash) was
requested, it would redirect to the same URL with the trailing slash.
This in turn would then be matched against the credsRegex and therefore
the metaproxy would be able to provide the correct role name in
response. We have observed that some instances no longer have this
redirect behavior. We know that there are some Go apps in our
infrastructure that have old versions of the AWS SDK that depended on
this redirect for requesting security credentials (unsure of apps in
other languages). While some are easy to upgrade, others are not as
straightforward so a catch all fix is to do the redirect behavior here
in the metaproxy rather than relying on AWS.

This is the commit that changed the behavior in the Go SDK, anything
compiled before this commit is affected, at least for Go apps:
https://github.com/aws/aws-sdk-go/commit/3cec29ed96b559eedd1d5f503804a940d7de21a7